[Cleanup][Ezra] separate data file
