+++ applied-mathematics.html
@@ -688,8 +688,9 @@
         });
       });
     </script>
-    <script src="javascript/course.js"></script>
-    <script>
+    <script type="module">
+      import { initializeProspectus } from './javascript/course/course.js';
+
       initializeProspectus('appmath');
     </script>
   </body>
