+++ computer-science.html
@@ -678,8 +678,9 @@
         });
       });
     </script>
-    <script src="javascript/course.js"></script>
-    <script>
+    <script type="module">
+      import { initializeProspectus } from './javascript/course/course.js';
+
       initializeProspectus('comsci');
     </script>
   </body>
